+++ brandonochiadotcom.spa/src/components/Layout.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Container } from 'reactstrap';
-import { NavMenu } from './NavMenu';
+import { NavMenu } from './navMenu/NavMenu';
 
 export class Layout extends Component {
   static displayName = Layout.name;
